+++ R/fps.R
@@ -213,6 +213,9 @@ fps <- function(lib, ...) {
 #' the reference library will be used.
 #' @param std The phase ID (e.g. "QUA.1") to be used as internal
 #' standard. Must match an ID provided in the \code{refs} parameter.
+#' @param force An optional string of phase ID's or names specifying which phases should be forced to
+#' remain throughout the automated full pattern summation. The ID's or names supplied must be present
+#' within the \code{lib$phases$phase_id} or \code{lib$phases$phase_name} columns.
 #' @param std_conc The concentration of the internal standard (if known) in weight percent. If
 #' unknown then use \code{std_conc = NA} (default), in which case it will be assumed that all phases sum
 #' to 100 percent.
@@ -328,7 +331,7 @@ fps <- function(lib, ...) {
 #' Eberl, D.D., 2003. User's guide to RockJock - A program for determining quantitative mineralogy from
 #' powder X-ray diffraction data. Boulder, CA.
 #' @export
-fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
+fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, force, std_conc,
                 tth_align, align, manual_align, tth_fps, shift,
                 remove_trace, ...) {
 
@@ -341,6 +344,30 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
 
     stop("Please make sure that there are no negative count intensities in the sample data",
          call. = FALSE)
+
+  }
+
+#Define force if missing
+  if (missing(force)) {
+
+    force <- c()
+
+  }
+
+#Extract phase ID's from force
+  if (length(force) > 0) {
+
+    wrong_force <- which(!force %in% lib$phases$phase_id & !force %in% lib$phases$phase_name)
+
+    if (length(wrong_force) > 0) {
+
+      stop(paste(c("\nThe following reference patterns specified in the force argument are not in the library:\n",
+                   paste(c(force[wrong_force]), collapse = ", "))),
+           call. = FALSE)
+
+    }
+
+    force <- lib$phases$phase_id[which(lib$phases$phase_id %in% force | lib$phases$phase_name %in% force)]
 
   }
 
@@ -532,7 +559,7 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
 
   if (length(wrong_spellings) > 0) {
 
-    stop(paste(c("\nThe following reference patterns are not in the library:\n",
+    stop(paste(c("\nThe following reference patterns specified in the refs argument are not in the library:\n",
                     paste(c(refs[wrong_spellings]), collapse = ", "))),
             call. = FALSE)
 
@@ -542,8 +569,8 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
 #END OF CONDITIONS, NOW SUBSET LIBRARY
 #-------------------------------------------------------------
 
-  #subset lib according to the refs vector
-  lib <- subset(lib, refs = refs, mode = "keep")
+  #subset lib according to the refs and force vector
+  lib <- subset(lib, refs = c(refs, force), mode = "keep")
 
   #Make sure that the phase identified as the internal standard is contained within the reference library
   if (!std == "none" & !std %in% lib$phases$phase_id) {
@@ -665,7 +692,7 @@ x <- o$par
 if (min(x) < 0) {
 
 remove_neg_out <- .remove_neg(x = x, lib = lib, smpl = smpl,
-                              solver = solver, obj = obj)
+                              solver = solver, obj = obj, force = force)
 
 x <- remove_neg_out[[1]]
 lib <- remove_neg_out[[2]]
@@ -675,7 +702,7 @@ lib <- remove_neg_out[[2]]
 } else {
 
   cat("\n-Applying non-negative least squares")
-  nnls_out <- .xrd_nnls(xrd.lib = lib, xrd.sample = smpl[, 2])
+  nnls_out <- .xrd_nnls(xrd.lib = lib, xrd.sample = smpl[, 2], force = force)
 
   lib$xrd <- nnls_out$xrd.lib
   x <- nnls_out$x
@@ -749,7 +776,7 @@ if(shift > 0 & length(x) > 1 & solver %in% c("Nelder-Mead", "BFGS", "CG")) {
 if (min(x) < 0) {
 
 remove_neg_out <- .remove_neg(x = x, lib = lib, smpl = smpl,
-                              solver = solver, obj = obj)
+                              solver = solver, obj = obj, force = force)
 
 x <- remove_neg_out[[1]]
 lib <- remove_neg_out[[2]]
@@ -832,6 +859,7 @@ inputs <- list("harmonise" = harmonise,
                "obj" = obj,
                "refs" = refs,
                "std" = std,
+               "force" = force,
                "std_conc" = std_conc,
                "tth_align" = tth_align,
                "align" = align,