+++ NEWS.md
@@ -1,6 +1,8 @@
 # powdR 1.2.1.999
 
 ## New features
+* `fps()` now accepts the `force` argument, forcing phases to remain in the final output even if their
+  coefficients are negative.
 * `fps()` and `afps()` will now stop if any of the phases specified in the `refs` argument are not
   in the library.
 * Warning messages now generally do not include the call.